requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.3
tensorflow==2.13.0
keras==2.13.0
statsmodels==0.14.0
prophet==1.1.5
scapy==2.5.0
psutil==5.9.5
watchdog==3.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
plotly==5.17.0
pyarrow==13.0.0
sqlalchemy==2.0.21
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyyaml==6.0.1
jsonschema==4.19.1
pydantic==2.4.2
cryptography==41.0.5
paramiko==3.3.1
slack-sdk==3.23.0
twilio==8.10.0
pillow==10.0.1
shap==0.42.3
lime==0.2.0
joblib==1.3.2
pytest==7.4.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.10.0
flake8==6.1.0
mypy==1.6.1
lxml==4.9.3
beautifulsoup4==4.12.2
gensim==4.3.1
imbalanced-learn==0.11.0
optuna==3.14.0
